well apparently I can just avoid the whole click garbage
Should be much more reliable now.

+++ archiver.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 import time
-from typing import List, Optional
+from typing import List, Optional, TypeVar
 from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
@@ -10,6 +10,7 @@ import argparse
 import json
 import os
 import requests
+import traceback
 
 
 @dataclass
@@ -17,14 +18,16 @@ class Post:
     url: str
     text: str
     images: List[str]
+    is_members: bool
+    relative_date: str
+    num_comments: Optional[str]
+    num_thumbs_up: Optional[str]
 
-    def record(self, output_dir: str):
+    def save(self, output_dir: str):
         id = self.url.split("/")[-1]
         dir = os.path.join(output_dir, id)
 
-        print(f"Trying to save `{id}` at `{dir}`...")
-
-        text_obj = {"url": self.url, "text": self.text, "image_urls": self.images}
+        # print(f"Trying to save `{id}` at `{dir}`...")
 
         if not os.path.exists(dir):
             try:
@@ -33,24 +36,50 @@ class Post:
                 print(f"err: couldn't make directory at {dir}")
                 return
 
-        data_path = os.path.join(dir, "post.txt")
-        with open(data_path, "w", encoding="utf-8") as f:
-            json.dump(text_obj, f, ensure_ascii=False, indent=4)
+        try:
+            data_path = os.path.join(dir, "post.txt")
+            with open(data_path, "w", encoding="utf-8") as f:
+                json.dump(self.__dict__, f, ensure_ascii=False, indent=4)
+        except:
+            print(f"err: couldn't save data dump at {data_path}")
 
         for itx, image in enumerate(self.images):
-            img_data = requests.get(image).content
-            img_name = f"{id}-{itx}.png"
-            img_path = os.path.join(dir, img_name)
-            with open(img_path, "wb") as f:
-                f.write(img_data)
+            try:
+                img_name = f"{id}-{itx}.png"
+                img_path = os.path.join(dir, img_name)
+
+                if not os.path.exists(img_path):
+                    img_data = requests.get(image).content
+                    with open(img_path, "wb") as f:
+                        f.write(img_data)
+                else:
+                    # print(f"Skipping saving image at {img_path} as it's already been saved.")
+                    pass
+            except:
+                print(f"err: couldn't save image `{image}` dump at {img_path}")
 
 
 class Archiver:
-    def __init__(self, url: str, output_dir: Optional[str]) -> None:
+    def __init__(
+        self,
+        url: str,
+        output_dir: Optional[str],
+        profile_dir: Optional[str] = None,
+    ) -> None:
         options = webdriver.ChromeOptions()
-        options.add_argument("--headless")
         options.add_argument("--window-size=1920,1080")
         options.add_argument("--start-maximized")
+
+        if profile_dir is not None:
+            split_dir = profile_dir.rsplit("/", 1)
+            user_data_dir = split_dir[0]
+            profile_dir = split_dir[1]
+
+            options.add_argument(f"--user-data-dir={user_data_dir}")
+            options.add_argument(f"--profile-directory={profile_dir}")
+        else:
+            # For some reason this doesn't work if I set a profile...????
+            options.add_argument("--headless")
 
         # Make sure the output directory exists... if not, then try and make it.
         output_dir = "archive-output" if output_dir is None else output_dir
@@ -75,37 +104,56 @@ class Archiver:
         return posts
 
     def handle_post(self, post: WebElement) -> None:
-        url = next(filter(self.filter_post_href, post.find_elements(By.TAG_NAME, "a")), None).get_attribute("href")
+        post_link = next(filter(self.filter_post_href, post.find_elements(By.TAG_NAME, "a")), None)
+
+        url = post_link.get_attribute("href")
         if url in self.seen:
             return
 
-        # Hit "Read more" for any visible posts.
-        read_more_buttons = post.find_elements(By.CLASS_NAME, "more-button")
-        if len(read_more_buttons) > 0:
-            read_more_button = read_more_buttons[0]
-            if read_more_button.is_displayed() and read_more_button.is_enabled():
-                read_more_button.click()
-            else:
-                return
+        relative_date = post_link.text
 
-        # Get the URL, the text, and any images
-        text = post.find_element(By.ID, "content").text
+        is_members = bool(post.find_elements(By.CLASS_NAME, "ytd-sponsors-only-badge-renderer"))
+
         images = [
-            url.split("=")[0] + "=s1080"
+            url.split("=")[0] + "=s3840"
             for url in filter(
                 lambda img: img is not None,
                 (img.get_attribute("src") for img in post.find_elements(By.TAG_NAME, "img")),
             )
         ]
 
+        comment_elements = post.find_elements(By.CSS_SELECTOR, "[aria-label=Comment]")
+        num_comments = comment_elements[0].text if comment_elements else None
+
+        thumbs_elements = post.find_elements(By.ID, "vote-count-middle")
+        num_thumbs_up = thumbs_elements[0].text if thumbs_elements else None
+
+        # This may need retries due to the above click.
+        try:
+            text_elements = post.find_elements(By.ID, "content")
+            if not text_elements:
+                return
+        except:
+            pass
+
+        text = text_elements[0].get_attribute("innerText")
+
         # We skip the first image since that's always the profile picture.
-        post = Post(url=url, text=text, images=images[1:])
+        post = Post(
+            url=url,
+            text=text,
+            images=images[1:],
+            is_members=is_members,
+            relative_date=relative_date,
+            num_comments=num_comments,
+            num_thumbs_up=num_thumbs_up,
+        )
         self.seen.add(url)
-        post.record(self.output_dir)
+        post.save(self.output_dir)
 
     def scrape(self):
         LOAD_SLEEP_SECS = 3
-        MAX_SAME_SEEN = 30
+        MAX_SAME_SEEN = 60
 
         try:
             self.driver.get(self.url)
@@ -115,11 +163,14 @@ class Archiver:
             same_seen = 0
 
             while True:
-                posts = self.find_posts()
-                for post in posts:
-                    self.handle_post(post)
+                try:
+                    posts = self.find_posts()
+                    for post in posts:
+                        self.handle_post(post)
+                except:
+                    continue
 
-                self.driver.execute_script("window.scrollBy(0, 500);")
+                self.driver.execute_script("window.scrollBy(0, 250);")
                 time.sleep(LOAD_SLEEP_SECS)
 
                 new_seen = len(self.seen)
@@ -133,8 +184,9 @@ class Archiver:
 
                 num_seen = new_seen
 
-        except Exception as err:
-            print(f"Encountered an error: {err}")
+        except Exception:
+            print("Encountered a fatal error:")
+            traceback.print_exc()
 
     def __enter__(self):
         return self
@@ -145,16 +197,25 @@ class Archiver:
 
 def main():
     parser = argparse.ArgumentParser(description="Archives YouTube community posts.")
-    parser.add_argument("url", type=str, help="The URL to grab posts from.")
+    parser.add_argument("url", type=str, help="The URL to try and grab posts from.")
     parser.add_argument("-o", "--output_dir", type=str, required=False, help="The directory to save to.")
+    parser.add_argument("-p", "--profile_dir", type=str, required=False, help="The path to an existing Chrome profile.")
+    parser.add_argument(
+        "-r", "--rerun", type=str, required=False, help="How many times to rerun the archiver. Should be at least 1."
+    )
 
     args = parser.parse_args()
     url = args.url
     output_dir = args.output_dir
+    profile_dir = args.profile_dir
+    rerun = int(args.rerun) if args.rerun and int(args.rerun) > 0 else 1
 
     try:
-        with Archiver(url, output_dir) as archiver:
-            archiver.scrape()
+        print(f"Running the archiver {rerun} time(s).")
+        for i in range(rerun):
+            print(f"===== Run {i + 1} ======")
+            with Archiver(url, output_dir, profile_dir) as archiver:
+                archiver.scrape()
     finally:
         print("Done!")
